requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==7.0.5
Django==4.1.4
django-click==2.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
pytz==2022.7.1
requests==2.28.2
simplejson==3.18.1
sqlparse==0.4.3
typer==0.7.0
uritemplate==4.1.1
urllib3==1.26.14
